fix(examples): widen RTC/LPM seed enums so the year base fits
The magic-number remediation folded the 2000 calendar epoch base into a
`enum : uint8_t` alongside the single-byte seed fields, which overflows
the underlying type ("enumerator value outside the range of underlying
type") and broke the cross-build of rtc_alarm, rtc_periodic_demo and the
four lpm standby demos. Promote those seed enums to `uint16_t`.

+++ examples/ek_ra8d2/hw_validated/hil/rtc_alarm/main.c
@@ -55,7 +55,7 @@ typedef enum : uint8_t {
 } rtc_demo_calendar_t;
 
 /** @brief Alarm offset from the current RTC reading. */
-typedef enum : uint8_t {
+typedef enum : uint16_t {
   k_rtc_demo_alarm_offset_s = 5U,
   k_rtc_demo_seed_year_lo   = 26U,   /**< 2026 - 2000. */
   k_rtc_demo_year_base      = 2000U, /**< Calendar epoch base for the year field. */
